+++ src/main/java/ru/xrm/app/transformers/VacancySalaryPropertyTransformer.java
@@ -3,10 +3,10 @@ package ru.xrm.app.transformers;
 public class VacancySalaryPropertyTransformer implements PropertyTransformer {
 
 	public Object transform(String from) {
-		Integer result=-1;
+		Long result=-1L;
 		String[] parts=from.split(" ");
 		if (parts.length==2){
-			result=Integer.valueOf(parts[0]);
+			result=Long.valueOf(parts[0]);
 		}
 		return result;
 	}
